requirements.txt
accelerate==0.21.0
bitsandbytes==0.41.0
deepspeed==0.9.5
einops==0.6.1
einops-exts==0.0.4
gradio==3.35.2
gradio-client==0.2.9
httpx==0.24.0
markdown2==2.4.10
peft==0.4.0
scikit-learn==1.2.2
semantic-version==2.10.0
sentencepiece==0.1.99
shortuuid==1.0.11
timm==0.6.13
tokenizers==0.13.3
--extra-index-url https://download.pytorch.org/whl/nightly/cu122
torch==2.0.1
--extra-index-url https://download.pytorch.org/whl/nightly/cu122
torchvision==0.15.2
transformers==4.31.0
wandb==0.16.0
wavedrom==2.0.3.post3
Pygments==2.16.1
streamlit==1.27.2
streamlit-image-coordinates==0.1.6
imutils==0.5.4
Flask==3.0.0
streamlit-dimensions==0.0.1
streamlit-cropper==0.2.2
label-studio
label-studio-sdk
opencv-python==4.8.1.78
seaborn==0.13.0
tensorboard==2.14.0
thop==0.1.1.post2209072238
azure-cognitiveservices-vision-computervision
shapely